requirements.txt
# Minimal requirements - extend as needed
jupyter
nbformat
google-api-python-client
openai
pandas
numpy
scikit-learn
matplotlib
seaborn
streamlit
google-generativeai
requests